Adicionado response/classe service/atualizado acesso controller

+++ src/main/java/com/recipePrice/recipePrice/dtos/ProductRequest.java
@@ -2,5 +2,5 @@ package com.recipePrice.recipePrice.dtos;
 
 public record ProductRequest(String name,
                              String unitOfMeasurement,
-                             Double price) {
+                             Float price) {
 }
